move a bunch of things around
