+++ origins/universe.py
@@ -4,159 +4,8 @@ import statistics
 import time
 import networkx as nx
 
-
-def distance(atom1, atom2):
-    """
-    Returns the distance between two atoms
-    """
-    x = atom1.x - atom2.x
-    y = atom1.y - atom2.y
-    return (x**2 + y**2) ** 0.5
-
-
-def components(magnitude, atom1, atom2):
-    """
-    Returns the x and y components of the force of atom2 on atom1
-    """
-    dist = max(distance(atom1, atom2), 1e-9)
-    xratio = (atom1.x - atom2.x) / dist
-    yratio = (atom1.y - atom2.y) / dist
-    fx = magnitude * xratio
-    fy = magnitude * yratio
-    return fx, fy
-
-
-class Force:
-    """
-    This represents the base class for all of the forces in the universe
-    """
-
-    def __init__(self, multiplier):
-        self.multiplier = multiplier
-
-    def apply(self, atoms):
-        raise NotImplementedError("This is a base class, this method must be overridden in the derived class.")
-
-
-class Wind(Force):
-    """
-    This represents an extended class for a basic force to replicate wind.
-    """
-
-    def apply(self, atoms):
-        for atom in atoms:
-            atom.x += self.multiplier
-
-
-class Electric(Force):
-    """ "
-    The definition of an electrical force between atoms with charges.
-    """
-
-    def force_components(self, atom1, atom2):
-        """
-        Returns the x and y components of the force of atom2 on atom1.
-        """
-        dist = max(distance(atom1, atom2), 1e-9)
-        magnitude = (self.multiplier * atom1.charge * atom2.charge) / dist**2
-        xratio = (atom1.x - atom2.x) / dist
-        yratio = (atom1.y - atom2.y) / dist
-
-        fx = magnitude * xratio
-        fy = magnitude * yratio
-        return fx, fy
-
-    def apply(self, atoms):
-        """
-        This finds and applies the electrical force between all combinations of 2 atoms in the universe"
-        """
-        for atom1, atom2 in combinations(atoms, 2):
-            fx, fy = self.force_components(atom1, atom2)
-            atom1.vx += fx / atom1.mass
-            atom1.vy += fy / atom1.mass
-
-            fx, fy = self.force_components(atom2, atom1)
-            atom2.vx += fx / atom2.mass
-            atom2.vy += fy / atom2.mass
-
-
-class Exclusion(Force):
-    """
-    The definition of a force due to atoms becoming too close to each other
-    """
-
-    def force_components(self, atom1, atom2):
-        """
-        Returns the x and y components of the force of atom2 on atom1.
-        """
-        dist = max(distance(atom1, atom2), 1e-9)
-        magnitude = (-1 * self.multiplier * atom1.charge * atom2.charge) / (dist*10)**3
-        xratio = (atom1.x - atom2.x) / dist
-        yratio = (atom1.y - atom2.y) / dist
-
-        fx = magnitude * xratio
-        fy = magnitude * yratio
-        return fx, fy
-
-    def apply(self, atoms):
-        """
-        This finds and applies the electrical force between all combinations of 2 atoms in the universe"
-        """
-        for atom1, atom2 in combinations(atoms, 2):
-            fx, fy = self.force_components(atom1, atom2)
-            atom1.vx += fx / atom1.mass
-            atom1.vy += fy / atom1.mass
-
-            fx, fy = self.force_components(atom2, atom1)
-            atom2.vx += fx / atom2.mass
-            atom2.vy += fy / atom2.mass
-
-
-class Atom:
-    """
-    The most basic object in the universe.
-    """
-
-    def __init__(self, x=0, y=0, vx=0, vy=0, mass=0, name="atom", max_x=25, max_y=25):
-        self.name = name
-        # Maximum positions, velocities, and masses of all atoms when randomly generating atoms
-        self.max_x = max_x
-        self.max_y = max_y
-        self.max_vx = 10
-        self.max_vy = 10
-        self.max_mass = 50
-        self.x = x
-        self.y = y
-        self.vx = vx
-        self.vy = vy
-        self.mass = mass
-
-    def randomize(self):
-        """
-        This randomly generates atom's positions, velcoties, and masses
-        """
-        self.x = random.random() * self.max_x
-        self.y = random.random() * self.max_y
-        self.vx = random.random() * self.max_vx
-        self.vy = random.random() * self.max_vy
-        self.mass = random.random() * self.max_mass
-
-
-class Ion(Atom):
-    """
-    An extended class of atoms that has charges.
-    """
-
-    def __init__(self, x=0, y=0, vx=0, vy=0, mass=0, charge=0, name="ion", *args, **kwargs):
-        # Maximum and minimum charges of ions randomly generated
-        self.max_charge = 2
-        self.min_charge = -2
-        self.charge = charge
-        super().__init__(x, y, vx, vy, mass, name=name, *args, **kwargs)
-
-    def randomize(self):
-        self.charge = random.randint(self.min_charge, self.max_charge)
-        super().randomize()
+from origins.forces import Electric
+from origins.calculations import distance, components
 
 
 class Universe:
